+++ src/components/conversationSetting/index.tsx
@@ -1,13 +1,15 @@
 import React from "react";
 import { useTranslations } from "next-intl";
-import { Divider, Input, Modal, Select } from "@/components/ui";
+import Icon from "@/components/icon";
+import { Divider, Input, Modal, Select, Tooltip } from "@/components/ui";
 import { useChannel, useLLM, useModel, useConversationSetting } from "@/hooks";
-import { cn } from "@/lib";
+import { cn, isUndefined } from "@/lib";
 
 interface IConversationSettings {
   name: string;
   model_type: string;
   model_value: string;
+  context_length: number;
 }
 
 const renderLabel = (item: any) => {
@@ -38,8 +40,13 @@ const renderModelLabel = (item: any) => {
   );
 };
 
+const lengthOptions = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10].map((e) => ({
+  label: String(e),
+  value: e,
+}));
+
 const ChangeTitle: React.FC = () => {
-  const t = useTranslations("nav");
+  const tChat = useTranslations("chat");
   const tCommon = useTranslations("common");
   const tPrompt = useTranslations("prompt");
 
@@ -47,13 +54,13 @@ const ChangeTitle: React.FC = () => {
   const { updateType, updateName } = useModel();
   const [open, setOpen] = useConversationSetting();
   const { openai, azure } = useLLM();
-  // const [open, setOpen] = React.useState(false);
   const LLMOptions = React.useMemo(() => [openai, azure], [openai, azure]);
   const [modelOptions, setModelOptions] = React.useState<any[]>([]);
   const [formData, setFormData] = React.useState<IConversationSettings>({
     name: "",
     model_type: "",
     model_value: "",
+    context_length: 8,
   });
 
   const onClose = () => setOpen(false);
@@ -69,18 +76,10 @@ const ChangeTitle: React.FC = () => {
     getModelOptions(value);
   };
 
-  const onChangeModelValue = (value: string) => {
-    setFormData((data) => {
-      const newData = JSON.parse(JSON.stringify(data)) as IConversationSettings;
-      newData.model_value = value;
-      return newData;
-    });
-  };
-
-  const onChangeName = (value: string) => {
-    setFormData((data) => {
-      const newData = JSON.parse(JSON.stringify(data)) as IConversationSettings;
-      newData.name = value;
+  const onChangeForm = (value: any, key: keyof IConversationSettings) => {
+    setFormData((data: IConversationSettings) => {
+      const newData = JSON.parse(JSON.stringify(data));
+      newData[key] = value;
       return newData;
     });
   };
@@ -94,11 +93,12 @@ const ChangeTitle: React.FC = () => {
   const submit = () => {
     setChannel((channel) => {
       const { activeId, list } = channel;
-      const findChannel = list.find((item) => item.channel_id === activeId);
-      if (!findChannel) return channel;
-      findChannel.channel_name = formData.name;
-      findChannel.channel_model.type = formData.model_type;
-      findChannel.channel_model.name = formData.model_value;
+      const findCh = list.find((item) => item.channel_id === activeId);
+      if (!findCh) return channel;
+      findCh.channel_name = formData.name;
+      findCh.channel_model.type = formData.model_type;
+      findCh.channel_model.name = formData.model_value;
+      findCh.channel_context_length = formData.context_length;
       updateType(formData.model_type);
       updateName(formData.model_value);
       return channel;
@@ -109,21 +109,30 @@ const ChangeTitle: React.FC = () => {
   React.useEffect(() => {
     if (open) {
       const { activeId, list } = channel;
-      const findChannel = list.find((item) => item.channel_id === activeId);
-      if (!findChannel) return;
+      const findCh = list.find((item) => item.channel_id === activeId);
+      if (!findCh) return;
+
+      const { channel_name, channel_model, channel_context_length } = findCh;
+
+      let context_length = 8;
+      if (!isUndefined(channel_context_length)) {
+        context_length = channel_context_length;
+      }
 
       setFormData({
-        name: findChannel.channel_name,
-        model_type: findChannel.channel_model.type,
-        model_value: findChannel.channel_model.name,
+        name: channel_name,
+        model_type: channel_model.type,
+        model_value: channel_model.name,
+        // default is 8
+        context_length,
       });
-      getModelOptions(findChannel.channel_model.type);
+      getModelOptions(findCh.channel_model.type);
     }
   }, [open]);
 
   return (
     <Modal
-      title={t("conversation-setting")}
+      title={tChat("conversation-setting")}
       maskClosable={false}
       open={open}
       onClose={onClose}
@@ -139,7 +148,7 @@ const ChangeTitle: React.FC = () => {
           onChange={onChangeModelType}
         />
         <div className="flex mt-2 items-center">
-          <div className="text-sm text-black/90 w-16 dark:text-white/90">
+          <div className="text-sm text-black/90 w-36 dark:text-white/90">
             {tPrompt("model")}
           </div>
           <div className="flex-1">
@@ -148,24 +157,40 @@ const ChangeTitle: React.FC = () => {
               options={modelOptions}
               renderLabel={renderModelLabel}
               value={formData.model_value}
-              onChange={onChangeModelValue}
+              onChange={(value) => onChangeForm(value, "model_value")}
             />
           </div>
         </div>
       </div>
       <Divider />
-      <div className="flex items-center">
-        <div className="text-sm text-black/90 w-16  dark:text-white/90">
-          {t("title")}
+      <div className="flex flex-col gap-2">
+        <div className="flex items-center">
+          <div className="text-sm text-black/90 w-36  dark:text-white/90">
+            {tChat("title")}
+          </div>
+          <Input
+            className="flex-1"
+            allowClear
+            maxLength={30}
+            placeholder={tCommon("please-enter") as string}
+            value={formData.name}
+            onChange={(value) => onChangeForm(value, "name")}
+          />
         </div>
-        <Input
-          className="flex-1"
-          allowClear
-          maxLength={30}
-          placeholder={tCommon("please-enter") as string}
-          value={formData.name}
-          onChange={onChangeName}
-        />
+        <div className="flex items-center">
+          <div className="text-sm text-black/90 w-36 dark:text-white/90 flex items-center gap-1">
+            <span>{tChat("context-length")}</span>
+            <Tooltip title={tChat("context-length-tip")}>
+              <Icon icon="question_line" size={18} />
+            </Tooltip>
+          </div>
+          <Select
+            className="flex-1"
+            options={lengthOptions}
+            value={formData.context_length}
+            onChange={(value) => onChangeForm(value, "context_length")}
+          />
+        </div>
       </div>
     </Modal>
   );